+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* `colsconfig-dialog-content` component:
+    * Fix sorting dropdown.
 
 ## [3.0.0-beta.1] - 2018-12-04
 ### Fixed